+++ R/bgm.R
@@ -92,7 +92,7 @@
 #' network is modeled with binary indicator variables that capture the structure
 #' of the network. The argument \code{edge_prior} is used to set a prior
 #' distribution for the edge indicator variables, i.e., the structure of the
-#' network. Currently, two options are implemented: The Bernoulli model
+#' network. Currently, three options are implemented: The Bernoulli model
 #' \code{edge_prior = "Bernoulli"} assumes that the probability that an edge
 #' between two variables is included is equal to \code{inclusion_probability}
 #' and independent of other edges or variables. When
@@ -102,8 +102,16 @@
 #' inclusion probability with shape parameters \code{beta_bernoulli_alpha} and
 #' \code{beta_bernoulli_beta}. If \code{beta_bernoulli_alpha = 1} and
 #' \code{beta_bernoulli_beta = 1}, this means that networks with the same
-#' complexity (number of edges) get the same prior weight. The default is
-#' \code{edge_prior = "Bernoulli"}.
+#' complexity (number of edges) get the same prior weight. The Stochastic Block
+#' model \code{edge_prior = "Stochastic Block"} assumes that nodes can be
+#' organized into blocks or clusters. In principle, the assignment of nodes to
+#' such clusters is unknown, and the model as implemented here considers all
+#' possible options (i.e., specifies a Dirichlet process on the node to block
+#' allocation). This model is advantageous when nodes are expected to fall into
+#' distinct clusters. The inclusion probabilities for the edges are defined at
+#' the level of the clusters, with a beta prior for the unknown inclusion
+#' probability with shape parameters \code{beta_bernoulli_alpha} and
+#' \code{beta_bernoulli_beta}. The default is \code{edge_prior = "Bernoulli"}.
 #' @param inclusion_probability The prior edge inclusion probability for the
 #' Bernoulli model. Can be a single probability, or a matrix of \code{p} rows
 #' and \code{p} columns specifying an inclusion probability for each edge pair.
@@ -112,6 +120,8 @@
 #' the Beta prior density for the Bernoulli inclusion probability. Must be
 #' positive numbers. Defaults to \code{beta_bernoulli_alpha = 1} and
 #' \code{beta_bernoulli_beta = 1}.
+#' @param dirichlet_alpha The shape of the Dirichlet prior on the node-to-block
+#' allocation parameters for the Stochastic Block model.
 #' @param na.action How do you want the function to handle missing data? If
 #' \code{na.action = "listwise"}, listwise deletion is used. If
 #' \code{na.action = "impute"}, missing data are imputed iteratively during the
@@ -253,10 +263,11 @@ bgm = function(x,
                threshold_alpha = 0.5,
                threshold_beta = 0.5,
                edge_selection = TRUE,
-               edge_prior = c("Bernoulli", "Beta-Bernoulli"),
+               edge_prior = c("Bernoulli", "Beta-Bernoulli", "Stochastic Block"),
                inclusion_probability = 0.5,
                beta_bernoulli_alpha = 1,
                beta_bernoulli_beta = 1,
+               dirichlet_alpha = 1,
                na.action = c("listwise", "impute"),
                save = FALSE,
                display_progress = TRUE) {
@@ -282,7 +293,8 @@ bgm = function(x,
                       edge_prior = edge_prior,
                       inclusion_probability = inclusion_probability,
                       beta_bernoulli_alpha = beta_bernoulli_alpha,
-                      beta_bernoulli_beta = beta_bernoulli_beta)
+                      beta_bernoulli_beta = beta_bernoulli_beta,
+                      dirichlet_alpha = dirichlet_alpha)
 
   # ----------------------------------------------------------------------------
   # The vector variable_type is now coded as boolean.
@@ -407,6 +419,7 @@ bgm = function(x,
                       theta = theta,
                       beta_bernoulli_alpha = beta_bernoulli_alpha,
                       beta_bernoulli_beta = beta_bernoulli_beta,
+                      dirichlet_alpha = dirichlet_alpha,
                       Index = Index,
                       iter = iter,
                       burnin = burnin,